+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExtendedBucket now supports creating origins for CloudFront.
 - The deploy method on \*Bucket types no longer accepts a scope.
 - Dead letter queues now default to 14 days of retention.
+- The aws-cdk-lib dependency has been updated to 2.232.1
+- Updated dependencies
 
 ### Removed