+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dead letter queues now default to 14 days of retention.
 - The aws-cdk-lib dependency has been updated to 2.232.1
 - Updated dependencies
+- Moved to using .prettierrc file
 
 ### Removed