+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # TrueMark CDK Library Changelog
+
+## 1.21.1
+### Patch Changes
+
+- 6f95456: Added include to BucketDeploymentConfig
+- a412028: Added @changesets/cli and changed to new publishing workflow